+++ pkg/project/project.go
@@ -6,7 +6,7 @@ var (
 	gitSHA        = "n/a"
 	name          = "cluster-operator"
 	source        = "https://github.com/giantswarm/cluster-operator"
-	version       = "0.23.13-dev"
+	version       = "0.23.13"
 )
 
 func BundleVersion() string {